Deduplicate metrics dependencies (#1561)
## Motivation

This PR is motivated by the regression identified in https://github.com/ZcashFoundation/zebra/issues/1349. That PR notes that the metrics stopped working for most of the crates other than `zebrad`.

## Solution

This PR resolves the regression by deduplicating the `metrics` crate dependency. During a recent change we upgraded the metrics version in `zebrad` and a couple other of our crates, but we never updated the dependencies in `zebra-state`, `zebra-consensus`, or `zebra-network`. This caused the metrics macros to attempt to retrieve the current metrics exporter through the wrong function. We would install the metrics exporter in `0.13`, but then attempt to look it up through the `0.12` crate, which contains a different instance of the metrics exporter static variable which is unset. Doing this causes the metrics macros to return `None` for the current exporter after which they just silently give up.

## Related Issues

closes https://github.com/ZcashFoundation/zebra/issues/1349

## Follow Up Work

I noticed we have quite a few duplicate dependencies in our tree. We might be able to save some compilation time by auditing those and deduplicating them as much as possible.

- https://github.com/ZcashFoundation/zebra/issues/1582

+++ zebra-network/src/peer_set/candidate_set.rs
@@ -149,9 +149,9 @@ where
     }
 
     pub fn next(&mut self) -> Option<MetaAddr> {
-        metrics::gauge!("candidate_set.disconnected", self.disconnected.len() as i64);
-        metrics::gauge!("candidate_set.gossiped", self.gossiped.len() as i64);
-        metrics::gauge!("candidate_set.failed", self.failed.len() as i64);
+        metrics::gauge!("candidate_set.disconnected", self.disconnected.len() as f64);
+        metrics::gauge!("candidate_set.gossiped", self.gossiped.len() as f64);
+        metrics::gauge!("candidate_set.failed", self.failed.len() as f64);
         debug!(
             disconnected_peers = self.disconnected.len(),
             gossiped_peers = self.gossiped.len(),
